Inline quaternion and Euler math in the PPO hot path
cProfile of a 2M-timestep stage-1 run showed
jax.scipy.spatial.transform.Rotation's np.vectorize wrapper consuming
~61% of wall time (97 s of 275 s in from_quat/as_matrix inside
build_actor_obs, plus ~70 s in from_matrix/as_euler inside
raw_to_env_action). The GPU sat at 1-5% SM utilization during
steady-state training.

Replace both calls with explicit JAX expressions:

- obs._quat_to_matrix: 10-mul / 6-add closed form for the unit
  quaternion -> rotation matrix.
- policy._matrix_to_euler_xyz: rotation matrix -> extrinsic xyz Euler
  angles matching scipy's convention, with the standard gimbal-lock
  branch absorbing the residual rotation into yaw when |cos pitch| is
  below GIMBAL_LOCK_EPS.

Per CLAUDE.md "Reimplement only when the library is incompatible with
our constraints" - the docstrings explain that the real-time training
budget forces the exemption.

Also fix the int32 checkpoint dtype warning. global_step and
total_timesteps were stored as int64, which JAX silently truncated to
int32 because JAX_ENABLE_X64 is off. Our maximum training budget (1e8)
fits comfortably in int32, so the explicit downgrade removes the
warning and the implicit overflow cast in the JAX literal cache.

+++ lsy_drone_racing/control/rl_song/train.py
@@ -840,9 +840,9 @@ def _save_checkpoint(
         "rng_key": train_state.rng_key,
         "env_rng_key": env.rng_key,
         "env_sim_rng_key": _env_sim_rng_key(env),
-        "global_step": jnp.asarray(train_state.global_step, dtype=jnp.int64),
+        "global_step": jnp.asarray(train_state.global_step, dtype=jnp.int32),
         "iteration": jnp.asarray(train_state.iteration, dtype=jnp.int32),
-        "total_timesteps": jnp.asarray(total_timesteps, dtype=jnp.int64),
+        "total_timesteps": jnp.asarray(total_timesteps, dtype=jnp.int32),
     }
     checkpointer = ocp.PyTreeCheckpointer()
     checkpointer.save(
